eliminar servicio modificado

+++ adquisiciones/src/main/java/com/giuct/adquisiciones/service/ServicioService.java
@@ -89,6 +89,16 @@ public class ServicioService extends AdquisicionService{
 
     @Override
     public void eliminarAdquisicion(Long id) {
-        servicioRepository.deleteById(id);
+        Optional<Servicio> servicioOptional = servicioRepository.findById(id);
+        if (servicioOptional.isPresent()){
+            Servicio servicio = servicioOptional.get();
+            FuenteFinanciamiento fuenteFinanciamiento = servicio.getFuenteFinanciamiento();
+            fuenteFinanciamiento.setMonto(fuenteFinanciamiento.getMonto()-servicio.getCosto());
+            servicioRepository.deleteById(id);
+            fuenteRepository.save(fuenteFinanciamiento);
+
+        }
+
+        throw new InvalidAdquisicionException("La adquisicion no existe");
     }
 }
